refactor(s3): split catalog and client boundaries
